refactor(config_reload): split out ConfigChangeset classifier
Extract ConfigChangeset struct, classify_config_change(), and its
five unit tests into a new src/app/config_changeset.rs module per
PR #408 review feedback. Keeps config_reload.rs focused on the
watcher loop and restart orchestration.

Result: config_reload.rs 668 -> 548 lines, 10 production functions
(at SRP threshold). All 10 affected tests still pass; clippy clean.
