+++ src/app/config_reload.rs
@@ -33,48 +33,9 @@ use tokio::task::JoinHandle;
 use tokio_util::sync::CancellationToken;
 use tracing::{info, warn};
 
+use crate::app::config_changeset::{ConfigChangeset, classify_config_change};
 use crate::app::{adapters, config_watcher, schedule, worker};
 use crate::config;
-
-// ─── Part B: config diff classifier ─────────────────────────────────────────
-
-/// What changed between two `UserConfig` snapshots.
-#[derive(Debug, Default, PartialEq)]
-pub struct ConfigChangeset {
-    /// Telegram routes, token, or admin_ids changed → adapter restart needed.
-    pub adapters_changed: bool,
-    /// Schedules changed → schedule_watcher restart needed.
-    pub schedules_changed: bool,
-    /// Sub-agent definitions changed → sub-agent worker restart needed.
-    pub sub_agents_changed: bool,
-    /// Only system_prompt changed (and nothing else requiring a restart).
-    pub system_prompt_only: bool,
-}
-
-/// Compare two `UserConfig` snapshots and classify what changed.
-/// Returns a `ConfigChangeset` that callers use to decide which components
-/// need to be restarted.
-pub fn classify_config_change(
-    old: &config::UserConfig,
-    new: &config::UserConfig,
-) -> ConfigChangeset {
-    let adapters_changed = old.telegram != new.telegram || old.discord != new.discord;
-    let schedules_changed = old.schedules != new.schedules;
-    let sub_agents_changed = old.agents != new.agents;
-    let system_prompt_changed = old.system_prompt != new.system_prompt;
-
-    // Classify as system-prompt-only when ONLY the system prompt differs and
-    // nothing that requires restarting any component has changed.
-    let system_prompt_only =
-        system_prompt_changed && !adapters_changed && !schedules_changed && !sub_agents_changed;
-
-    ConfigChangeset {
-        adapters_changed,
-        schedules_changed,
-        sub_agents_changed,
-        system_prompt_only,
-    }
-}
 
 // ─── Part A: AgentComponents with cooperative cancellation ──────────────────
 
@@ -516,87 +477,6 @@ fn file_mtime(path: &str) -> Option<std::time::SystemTime> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::config::{ScheduleAction, ScheduleDef, TelegramRoute, TelegramRoutesConfig};
-
-    fn base_config() -> config::UserConfig {
-        config::UserConfig {
-            system_prompt: "You are a helpful assistant.".into(),
-            ..Default::default()
-        }
-    }
-
-    // ── classify_config_change ────────────────────────────────────────────────
-
-    #[test]
-    fn test_classify_no_change() {
-        let cfg = base_config();
-        let cs = classify_config_change(&cfg, &cfg);
-        assert_eq!(cs, ConfigChangeset::default());
-        assert!(!cs.system_prompt_only);
-    }
-
-    #[test]
-    fn test_classify_system_prompt_only() {
-        let old = base_config();
-        let mut new = old.clone();
-        new.system_prompt = "New prompt.".into();
-        let cs = classify_config_change(&old, &new);
-        assert!(cs.system_prompt_only);
-        assert!(!cs.adapters_changed);
-        assert!(!cs.schedules_changed);
-        assert!(!cs.sub_agents_changed);
-    }
-
-    #[test]
-    fn test_classify_telegram_routes_changed() {
-        let old = base_config();
-        let mut new = old.clone();
-        new.telegram = Some(TelegramRoutesConfig {
-            routes: vec![TelegramRoute {
-                chat_id: 12345,
-                mention_only: false,
-                name: None,
-                route_to: None,
-            }],
-        });
-        let cs = classify_config_change(&old, &new);
-        assert!(cs.adapters_changed);
-        assert!(!cs.system_prompt_only);
-        assert!(!cs.schedules_changed);
-    }
-
-    #[test]
-    fn test_classify_schedules_changed() {
-        let old = base_config();
-        let mut new = old.clone();
-        new.schedules = vec![ScheduleDef {
-            cron: "0 9 * * *".into(),
-            target: "agent:dev".into(),
-            action: ScheduleAction::Raw,
-            config: None,
-            timezone: None,
-        }];
-        let cs = classify_config_change(&old, &new);
-        assert!(cs.schedules_changed);
-        assert!(!cs.adapters_changed);
-        assert!(!cs.system_prompt_only);
-    }
-
-    #[test]
-    fn test_classify_system_prompt_and_adapters_not_prompt_only() {
-        let old = base_config();
-        let mut new = old.clone();
-        new.system_prompt = "Updated.".into();
-        new.telegram = Some(TelegramRoutesConfig { routes: vec![] });
-        let cs = classify_config_change(&old, &new);
-        assert!(
-            !cs.system_prompt_only,
-            "both changed — should not be prompt-only"
-        );
-        assert!(cs.adapters_changed);
-    }
-
-    // ── AgentComponents helpers ──────────────────────────────────────────────
 
     #[test]
     fn test_agent_components_summary() {
